Update home logic

+++ app/src/main/java/com/delivery/delivery/model/Location.kt
@@ -1,7 +1,7 @@
 package com.delivery.delivery.model
 
 data class Location(
-    val lat: Long,
-    val Lng: Long,
+    val lat: Double,
+    val Lng: Double,
     val address: String
 )
